pyproject.toml
[project]
name = "delphi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.23.0",
    "chainlit>=2.5.5",
    "google-generativeai>=0.8.0",
    "openai>=1.84.0",
    "openai-agents>=0.0.17",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.13",
]
